+++ sdk/resourcemanager/datafactory/armdatafactory/models.go
@@ -2116,6 +2116,9 @@ type AzureBlobFSWriteSettings struct {
 
 	// The maximum concurrent connection count for the source data store. Type: integer (or Expression with resultType integer).
 	MaxConcurrentConnections any
+
+	// Specify the custom metadata to be added to sink data. Type: array of objects (or Expression with resultType array of objects).
+	Metadata []*MetadataItem
 }
 
 // GetStoreWriteSettings implements the StoreWriteSettingsClassification interface for type AzureBlobFSWriteSettings.
@@ -2125,6 +2128,7 @@ func (a *AzureBlobFSWriteSettings) GetStoreWriteSettings() *StoreWriteSettings {
 		CopyBehavior:             a.CopyBehavior,
 		DisableMetricsCollection: a.DisableMetricsCollection,
 		MaxConcurrentConnections: a.MaxConcurrentConnections,
+		Metadata:                 a.Metadata,
 		Type:                     a.Type,
 	}
 }
@@ -2323,6 +2327,9 @@ type AzureBlobStorageWriteSettings struct {
 
 	// The maximum concurrent connection count for the source data store. Type: integer (or Expression with resultType integer).
 	MaxConcurrentConnections any
+
+	// Specify the custom metadata to be added to sink data. Type: array of objects (or Expression with resultType array of objects).
+	Metadata []*MetadataItem
 }
 
 // GetStoreWriteSettings implements the StoreWriteSettingsClassification interface for type AzureBlobStorageWriteSettings.
@@ -2332,6 +2339,7 @@ func (a *AzureBlobStorageWriteSettings) GetStoreWriteSettings() *StoreWriteSetti
 		CopyBehavior:             a.CopyBehavior,
 		DisableMetricsCollection: a.DisableMetricsCollection,
 		MaxConcurrentConnections: a.MaxConcurrentConnections,
+		Metadata:                 a.Metadata,
 		Type:                     a.Type,
 	}
 }
@@ -3026,6 +3034,9 @@ type AzureDataLakeStoreWriteSettings struct {
 
 	// The maximum concurrent connection count for the source data store. Type: integer (or Expression with resultType integer).
 	MaxConcurrentConnections any
+
+	// Specify the custom metadata to be added to sink data. Type: array of objects (or Expression with resultType array of objects).
+	Metadata []*MetadataItem
 }
 
 // GetStoreWriteSettings implements the StoreWriteSettingsClassification interface for type AzureDataLakeStoreWriteSettings.
@@ -3035,6 +3046,7 @@ func (a *AzureDataLakeStoreWriteSettings) GetStoreWriteSettings() *StoreWriteSet
 		CopyBehavior:             a.CopyBehavior,
 		DisableMetricsCollection: a.DisableMetricsCollection,
 		MaxConcurrentConnections: a.MaxConcurrentConnections,
+		Metadata:                 a.Metadata,
 		Type:                     a.Type,
 	}
 }
@@ -3577,6 +3589,9 @@ type AzureFileStorageWriteSettings struct {
 
 	// The maximum concurrent connection count for the source data store. Type: integer (or Expression with resultType integer).
 	MaxConcurrentConnections any
+
+	// Specify the custom metadata to be added to sink data. Type: array of objects (or Expression with resultType array of objects).
+	Metadata []*MetadataItem
 }
 
 // GetStoreWriteSettings implements the StoreWriteSettingsClassification interface for type AzureFileStorageWriteSettings.
@@ -3586,6 +3601,7 @@ func (a *AzureFileStorageWriteSettings) GetStoreWriteSettings() *StoreWriteSetti
 		CopyBehavior:             a.CopyBehavior,
 		DisableMetricsCollection: a.DisableMetricsCollection,
 		MaxConcurrentConnections: a.MaxConcurrentConnections,
+		Metadata:                 a.Metadata,
 		Type:                     a.Type,
 	}
 }
@@ -3673,7 +3689,7 @@ type AzureFunctionActivityTypeProperties struct {
 	// Represents the headers that will be sent to the request. For example, to set the language and type on a request: "headers"
 	// : { "Accept-Language": "en-us", "Content-Type": "application/json" }. Type:
 	// string (or Expression with resultType string).
-	Headers any
+	Headers map[string]*string
 }
 
 // AzureFunctionLinkedService - Azure Function linked service.
@@ -3714,7 +3730,8 @@ func (a *AzureFunctionLinkedService) GetLinkedService() *LinkedService {
 
 // AzureFunctionLinkedServiceTypeProperties - Azure Function linked service properties.
 type AzureFunctionLinkedServiceTypeProperties struct {
-	// REQUIRED; The endpoint of the Azure Function App. URL will be in the format https://.azurewebsites.net.
+	// REQUIRED; The endpoint of the Azure Function App. URL will be in the format https://.azurewebsites.net. Type: string (or
+	// Expression with resultType string).
 	FunctionAppURL any
 
 	// Type of authentication (Required to specify MSI) used to connect to AzureFunction. Type: string (or Expression with resultType
@@ -3731,7 +3748,7 @@ type AzureFunctionLinkedServiceTypeProperties struct {
 	// Function or Host key for Azure Function App.
 	FunctionKey SecretBaseClassification
 
-	// Allowed token audiences for azure function.
+	// Allowed token audiences for azure function. Type: string (or Expression with resultType string).
 	ResourceID any
 }
 
@@ -3964,8 +3981,8 @@ type AzureMLExecutePipelineActivityTypeProperties struct {
 	ContinueOnStepFailure any
 
 	// Dictionary used for changing data path assignments without retraining. Values will be passed in the dataPathAssignments
-	// property of the published pipeline execution request. Type: object with key
-	// value pairs (or Expression with resultType object).
+	// property of the published pipeline execution request. Type: object (or
+	// Expression with resultType object).
 	DataPathAssignments any
 
 	// Run history experiment name of the pipeline run. This information will be passed in the ExperimentName property of the
@@ -5258,7 +5275,7 @@ type AzureSQLSource struct {
 	MaxConcurrentConnections any
 
 	// The partition mechanism that will be used for Sql read in parallel. Possible values include: "None", "PhysicalPartitionsOfTable",
-	// "DynamicRange".
+	// "DynamicRange". Type: string (or Expression with resultType string).
 	PartitionOption any
 
 	// The settings that will be leveraged for Sql source partitioning.
@@ -6546,10 +6563,10 @@ type CmdkeySetupTypeProperties struct {
 	// REQUIRED; The password of data source access.
 	Password SecretBaseClassification
 
-	// REQUIRED; The server name of data source access.
+	// REQUIRED; The server name of data source access. Type: string.
 	TargetName any
 
-	// REQUIRED; The user name of data source access.
+	// REQUIRED; The user name of data source access. Type: string.
 	UserName any
 }
 
@@ -11778,6 +11795,9 @@ type FileServerWriteSettings struct {
 
 	// The maximum concurrent connection count for the source data store. Type: integer (or Expression with resultType integer).
 	MaxConcurrentConnections any
+
+	// Specify the custom metadata to be added to sink data. Type: array of objects (or Expression with resultType array of objects).
+	Metadata []*MetadataItem
 }
 
 // GetStoreWriteSettings implements the StoreWriteSettingsClassification interface for type FileServerWriteSettings.
@@ -11787,6 +11807,7 @@ func (f *FileServerWriteSettings) GetStoreWriteSettings() *StoreWriteSettings {
 		CopyBehavior:             f.CopyBehavior,
 		DisableMetricsCollection: f.DisableMetricsCollection,
 		MaxConcurrentConnections: f.MaxConcurrentConnections,
+		Metadata:                 f.Metadata,
 		Type:                     f.Type,
 	}
 }
@@ -13713,8 +13734,7 @@ type HDInsightOnDemandLinkedServiceTypeProperties struct {
 	// REQUIRED; The resource group where the cluster belongs. Type: string (or Expression with resultType string).
 	ClusterResourceGroup any
 
-	// REQUIRED; Number of worker/data nodes in the cluster. Suggestion value: 4. Type: string (or Expression with resultType
-	// string).
+	// REQUIRED; Number of worker/data nodes in the cluster. Suggestion value: 4. Type: int (or Expression with resultType int).
 	ClusterSize any
 
 	// REQUIRED; The customer’s subscription to host the cluster. Type: string (or Expression with resultType string).
@@ -14226,8 +14246,8 @@ type HTTPLinkedServiceTypeProperties struct {
 	// string).
 	URL any
 
-	// The additional HTTP headers in the request to RESTful API used for authorization. Type: object (or Expression with resultType
-	// object).
+	// The additional HTTP headers in the request to RESTful API used for authorization. Type: key value pairs (value should be
+	// string type).
 	AuthHeaders any
 
 	// The authentication type to be used to connect to the HTTP server.
@@ -15822,6 +15842,9 @@ type JSONDataset struct {
 	// DatasetSchemaDataElement.
 	Schema any
 
+	// Anything
+	Schema1 any
+
 	// Columns that define the structure of the dataset. Type: array (or Expression with resultType array), itemType: DatasetDataElement.
 	Structure any
 
@@ -16526,6 +16549,9 @@ type LakeHouseWriteSettings struct {
 
 	// The maximum concurrent connection count for the source data store. Type: integer (or Expression with resultType integer).
 	MaxConcurrentConnections any
+
+	// Specify the custom metadata to be added to sink data. Type: array of objects (or Expression with resultType array of objects).
+	Metadata []*MetadataItem
 }
 
 // GetStoreWriteSettings implements the StoreWriteSettingsClassification interface for type LakeHouseWriteSettings.
@@ -16535,6 +16561,7 @@ func (l *LakeHouseWriteSettings) GetStoreWriteSettings() *StoreWriteSettings {
 		CopyBehavior:             l.CopyBehavior,
 		DisableMetricsCollection: l.DisableMetricsCollection,
 		MaxConcurrentConnections: l.MaxConcurrentConnections,
+		Metadata:                 l.Metadata,
 		Type:                     l.Type,
 	}
 }
@@ -19021,8 +19048,8 @@ type ODataLinkedServiceTypeProperties struct {
 	// Specify the credential type (key or cert) is used for service principal.
 	AADServicePrincipalCredentialType *ODataAADServicePrincipalCredentialType
 
-	// The additional HTTP headers in the request to RESTful API used for authorization. Type: object (or Expression with resultType
-	// object).
+	// The additional HTTP headers in the request to RESTful API used for authorization. Type: key value pairs (value should be
+	// string type).
 	AuthHeaders any
 
 	// Type of authentication used to connect to the OData service.
@@ -22552,7 +22579,7 @@ type RestSink struct {
 	// REQUIRED; Copy sink type.
 	Type *string
 
-	// The additional HTTP headers in the request to the RESTful API. Type: string (or Expression with resultType string).
+	// The additional HTTP headers in the request to the RESTful API. Type: key value pairs (value should be string type).
 	AdditionalHeaders any
 
 	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
@@ -22562,7 +22589,7 @@ type RestSink struct {
 	DisableMetricsCollection any
 
 	// Http Compression Type to Send data in compressed format with Optimal Compression Level, Default is None. And The Only Supported
-	// option is Gzip.
+	// option is Gzip. Type: string (or Expression with resultType string).
 	HTTPCompressionType any
 
 	// The timeout (TimeSpan) to get an HTTP response. It is the timeout to get a response, not the timeout to read response data.
@@ -22611,8 +22638,7 @@ type RestSource struct {
 	// REQUIRED; Copy source type.
 	Type *string
 
-	// Specifies the additional columns to be added to source data. Type: array of objects(AdditionalColumns) (or Expression with
-	// resultType array of objects).
+	// Specifies the additional columns to be added to source data. Type: key value pairs (value should be string type).
 	AdditionalColumns any
 
 	// The additional HTTP headers in the request to the RESTful API. Type: string (or Expression with resultType string).
@@ -22825,7 +22851,7 @@ type SQLDWSource struct {
 	MaxConcurrentConnections any
 
 	// The partition mechanism that will be used for Sql read in parallel. Possible values include: "None", "PhysicalPartitionsOfTable",
-	// "DynamicRange".
+	// "DynamicRange". Type: string (or Expression with resultType string).
 	PartitionOption any
 
 	// The settings that will be leveraged for Sql source partitioning.
@@ -22938,7 +22964,7 @@ type SQLMISink struct {
 	// Write batch timeout. Type: string (or Expression with resultType string), pattern: ((\d+).)?(\d\d):(60|([0-5][0-9])):(60|([0-5][0-9])).
 	WriteBatchTimeout any
 
-	// White behavior when copying data into azure SQL MI. Type: SqlWriteBehaviorEnum (or Expression with resultType SqlWriteBehaviorEnum)
+	// White behavior when copying data into azure SQL MI. Type: string (or Expression with resultType string)
 	WriteBehavior any
 }
 
@@ -22980,7 +23006,7 @@ type SQLMISource struct {
 	MaxConcurrentConnections any
 
 	// The partition mechanism that will be used for Sql read in parallel. Possible values include: "None", "PhysicalPartitionsOfTable",
-	// "DynamicRange".
+	// "DynamicRange". Type: string (or Expression with resultType string).
 	PartitionOption any
 
 	// The settings that will be leveraged for Sql source partitioning.
@@ -23159,7 +23185,7 @@ type SQLServerSink struct {
 	// Write batch timeout. Type: string (or Expression with resultType string), pattern: ((\d+).)?(\d\d):(60|([0-5][0-9])):(60|([0-5][0-9])).
 	WriteBatchTimeout any
 
-	// Write behavior when copying data into sql server. Type: SqlWriteBehaviorEnum (or Expression with resultType SqlWriteBehaviorEnum)
+	// Write behavior when copying data into sql server. Type: string (or Expression with resultType string).
 	WriteBehavior any
 }
 
@@ -23201,7 +23227,7 @@ type SQLServerSource struct {
 	MaxConcurrentConnections any
 
 	// The partition mechanism that will be used for Sql read in parallel. Possible values include: "None", "PhysicalPartitionsOfTable",
-	// "DynamicRange".
+	// "DynamicRange". Type: string (or Expression with resultType string).
 	PartitionOption any
 
 	// The settings that will be leveraged for Sql source partitioning.
@@ -23444,7 +23470,7 @@ type SQLSink struct {
 	// Write batch timeout. Type: string (or Expression with resultType string), pattern: ((\d+).)?(\d\d):(60|([0-5][0-9])):(60|([0-5][0-9])).
 	WriteBatchTimeout any
 
-	// Write behavior when copying data into sql. Type: SqlWriteBehaviorEnum (or Expression with resultType SqlWriteBehaviorEnum)
+	// Write behavior when copying data into sql. Type: string (or Expression with resultType string).
 	WriteBehavior any
 }
 
@@ -23486,7 +23512,7 @@ type SQLSource struct {
 	MaxConcurrentConnections any
 
 	// The partition mechanism that will be used for Sql read in parallel. Possible values include: "None", "PhysicalPartitionsOfTable",
-	// "DynamicRange".
+	// "DynamicRange". Type: string (or Expression with resultType string).
 	PartitionOption any
 
 	// The settings that will be leveraged for Sql source partitioning.
@@ -26642,6 +26668,9 @@ type SftpWriteSettings struct {
 	// The maximum concurrent connection count for the source data store. Type: integer (or Expression with resultType integer).
 	MaxConcurrentConnections any
 
+	// Specify the custom metadata to be added to sink data. Type: array of objects (or Expression with resultType array of objects).
+	Metadata []*MetadataItem
+
 	// Specifies the timeout for writing each chunk to SFTP server. Default value: 01:00:00 (one hour). Type: string (or Expression
 	// with resultType string).
 	OperationTimeout any
@@ -26658,6 +26687,7 @@ func (s *SftpWriteSettings) GetStoreWriteSettings() *StoreWriteSettings {
 		CopyBehavior:             s.CopyBehavior,
 		DisableMetricsCollection: s.DisableMetricsCollection,
 		MaxConcurrentConnections: s.MaxConcurrentConnections,
+		Metadata:                 s.Metadata,
 		Type:                     s.Type,
 	}
 }
@@ -27974,6 +28004,9 @@ type StoreWriteSettings struct {
 
 	// The maximum concurrent connection count for the source data store. Type: integer (or Expression with resultType integer).
 	MaxConcurrentConnections any
+
+	// Specify the custom metadata to be added to sink data. Type: array of objects (or Expression with resultType array of objects).
+	Metadata []*MetadataItem
 }
 
 // GetStoreWriteSettings implements the StoreWriteSettingsClassification interface for type StoreWriteSettings.
@@ -29765,13 +29798,23 @@ type WebActivityTypeProperties struct {
 	// When set to true, Certificate validation will be disabled.
 	DisableCertValidation *bool
 
+	// Timeout for the HTTP request to get a response. Format is in TimeSpan (hh:mm:ss). This value is the timeout to get a response,
+	// not the activity timeout. The default value is 00:01:00 (1 minute). The
+	// range is from 1 to 10 minutes
+	HTTPRequestTimeout any
+
 	// Represents the headers that will be sent to the request. For example, to set the language and type on a request: "headers"
 	// : { "Accept-Language": "en-us", "Content-Type": "application/json" }. Type:
 	// string (or Expression with resultType string).
-	Headers any
+	Headers map[string]*string
 
 	// List of linked services passed to web endpoint.
 	LinkedServices []*LinkedServiceReference
+
+	// Option to disable invoking HTTP GET on location given in response header of a HTTP 202 Response. If set true, it stops
+	// invoking HTTP GET on http location given in response header. If set false then
+	// continues to invoke HTTP GET call on location given in http response headers.
+	TurnOffAsync *bool
 }
 
 // WebAnonymousAuthentication - A WebLinkedService that uses anonymous authentication to communicate with an HTTP endpoint.
@@ -29922,7 +29965,7 @@ type WebHookActivityTypeProperties struct {
 	// Represents the headers that will be sent to the request. For example, to set the language and type on a request: "headers"
 	// : { "Accept-Language": "en-us", "Content-Type": "application/json" }. Type:
 	// string (or Expression with resultType string).
-	Headers any
+	Headers map[string]*string
 
 	// When set to true, statusCode, output and error in callback request body will be consumed by activity. The activity can
 	// be marked as failed by setting statusCode >= 400 in callback request. Default is